Move auth handling in MainViewModel
