+++ gnd/src/main/java/com/google/android/gnd/MainViewModel.java
@@ -22,14 +22,19 @@ import androidx.lifecycle.MutableLiveData;
 import com.google.android.gnd.model.Project;
 import com.google.android.gnd.repository.FeatureRepository;
 import com.google.android.gnd.repository.ProjectRepository;
+import com.google.android.gnd.repository.UserRepository;
 import com.google.android.gnd.rx.Loadable;
+import com.google.android.gnd.system.auth.AuthenticationManager;
+import com.google.android.gnd.system.auth.SignInState;
 import com.google.android.gnd.ui.common.AbstractViewModel;
+import com.google.android.gnd.ui.common.EphemeralPopups;
 import com.google.android.gnd.ui.common.Navigator;
 import com.google.android.gnd.ui.common.SharedViewModel;
 import com.google.android.gnd.ui.home.HomeScreenFragmentDirections;
 import com.google.android.gnd.ui.signin.SignInFragmentDirections;
 import io.reactivex.Completable;
 import javax.inject.Inject;
+import timber.log.Timber;
 
 /** Top-level view model representing state of the {@link MainActivity} shared by all fragments. */
 @SharedViewModel
@@ -37,18 +42,26 @@ public class MainViewModel extends AbstractViewModel {
 
   private final ProjectRepository projectRepository;
   private final FeatureRepository featureRepository;
+  private final UserRepository userRepository;
   private final Navigator navigator;
+  private EphemeralPopups popups;
+
   private MutableLiveData<WindowInsetsCompat> windowInsetsLiveData;
 
   @Inject
   public MainViewModel(
       ProjectRepository projectRepository,
       FeatureRepository featureRepository,
-      Navigator navigator) {
+      UserRepository userRepository,
+      Navigator navigator,
+      AuthenticationManager authenticationManager,
+      EphemeralPopups popups) {
     windowInsetsLiveData = new MutableLiveData<>();
     this.projectRepository = projectRepository;
     this.featureRepository = featureRepository;
+    this.userRepository = userRepository;
     this.navigator = navigator;
+    this.popups = popups;
 
     // TODO: Move to background service.
     disposeOnClear(
@@ -56,6 +69,8 @@ public class MainViewModel extends AbstractViewModel {
             .getActiveProjectOnceAndStream()
             .switchMapCompletable(this::syncFeatures)
             .subscribe());
+
+    disposeOnClear(authenticationManager.getSignInState().subscribe(this::onSignInStateChange));
   }
 
   /**
@@ -74,6 +89,38 @@ public class MainViewModel extends AbstractViewModel {
 
   void onApplyWindowInsets(WindowInsetsCompat insets) {
     windowInsetsLiveData.setValue(insets);
+  }
+
+  private void onSignInStateChange(SignInState signInState) {
+    Timber.d("Auth status change: %s", signInState.state());
+    switch (signInState.state()) {
+      case SIGNED_OUT:
+        // TODO: Check auth status whenever fragments resumes.
+        onSignedOut();
+        break;
+      case SIGNING_IN:
+        // TODO: Show/hide spinner.
+        break;
+      case SIGNED_IN:
+        signInState
+            .getUser()
+            .ifPresentOrElse(
+                user -> userRepository.saveUser(user).subscribe(() -> onSignedIn()),
+                () -> Timber.e("User signed in but missing"));
+        break;
+      case ERROR:
+        onSignInError(signInState);
+        break;
+      default:
+        Timber.e("Unhandled state: %s", signInState.state());
+        break;
+    }
+  }
+
+  private void onSignInError(SignInState signInState) {
+    Timber.d("Authentication error : %s", signInState.error());
+    popups.showError(R.string.sign_in_unsuccessful);
+    onSignedOut();
   }
 
   void onSignedOut() {
